+++ redis_enterprise/datadog_checks/redis_enterprise/check.py
@@ -18,7 +18,7 @@ class RedisEnterpriseCheck(OpenMetricsBaseCheckV2):
     def _parse_config(self):
         self.scraper_configs = []
         metrics_endpoint = self.instance.get('openmetrics_endpoint')
-        tls_verify = self.instance.get("tls_verify", False)
+        tls_verify = self.instance.get("tls_verify", True)
         metrics = self.get_default_config()
 
         additional = []